api park branch
api park branch

+++ BackEnd/park_management/src/main/java/park/management/com/vn/repository/ParkBranchRepository.java
@@ -4,6 +4,9 @@ import park.management.com.vn.entity.ParkBranch;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface ParkBranchRepository extends JpaRepository<ParkBranch,Long> {
+public interface ParkBranchRepository extends JpaRepository<ParkBranch,Integer> {
+
 }
